[WAF] fix attribute `ExpireTime` type for `resource/opentelekomcloud_waf_certificate_v1` (#2417)
[WAF] fix attribute `ExpireTime` type for `resource/opentelekomcloud_waf_certificate_v1`

Summary of the Pull Request
Small fix for int64 values
PR Checklist

 Refers to: #2412
 Tests added/passed.
 Documentation updated.
 Schema updated.
 Release notes added.

Acceptance Steps Performed
=== RUN   TestAccWafCertificateV1_basic
--- PASS: TestAccWafCertificateV1_basic (41.88s)
=== RUN   TestAccWafCertificateV1_validation
--- PASS: TestAccWafCertificateV1_validation (3.80s)
PASS

Process finished with the exit code 0

+++ opentelekomcloud/services/waf/resource_certificate_v1.go
@@ -108,8 +108,7 @@ func resourceWafCertificateV1Read(_ context.Context, d *schema.ResourceData, met
 		return fmterr.Errorf("error retrieving OpenTelekomCloud Waf Certificate: %w", err)
 	}
 
-	expires := time.Unix(int64(n.ExpireTime/1000), 0).UTC().Format("2006-01-02 15:04:05 MST")
-
+	expires := time.Unix(n.ExpireTime/1000, 0).UTC().Format("2006-01-02 15:04:05 MST")
 	mErr := multierror.Append(
 		d.Set("name", n.Name),
 		d.Set("expires", expires),
